requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Migrate
Flask-Bcrypt
Pillow
python-dotenv
mistune